+++ docs-src/pages/changelog.svelte
@@ -1,5 +1,5 @@
 <h1>Changelog</h1>
-<h2>v9.1.0 <em>(2023-09-26)</em></h2>
+<h2>v9.1.0 <em>(2023-09-27)</em></h2>
 <ul>
 <li>New components: <code>InputRating</code>, <code>Tag</code>, <code>InputTag</code>, <code>InputTime</code>.</li>
 <li>Add <code>hideTip</code> and more, to <code>Popover</code>.</li>